backend/requirements.txt

mysql-connector-python>=8.0
pandas>=1.0
python-dotenv>=1.0
bcrypt>=4.0
PyJWT>=2.0
requests>=2.0
pypdf>=3.17.0
pdfplumber
